+++ packages/bsky/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @atproto/bsky
 
+## 0.0.107
+
+### Patch Changes
+
+- Updated dependencies [[`0832a377d`](https://github.com/bluesky-social/atproto/commit/0832a377d269584a906d5062ebb5e2e6307f9c61)]:
+  - @atproto/xrpc-server@0.7.7
+  - @atproto-labs/xrpc-utils@0.0.3
+  - @atproto/sync@0.1.10
+
 ## 0.0.106
 
 ### Patch Changes