+++ packages/internal/xrpc-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atproto-labs/xrpc-utils
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`0832a377d`](https://github.com/bluesky-social/atproto/commit/0832a377d269584a906d5062ebb5e2e6307f9c61)]:
+  - @atproto/xrpc-server@0.7.7
+
 ## 0.0.2
 
 ### Patch Changes