+++ packages/ozone/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atproto/ozone
 
+## 0.1.68
+
+### Patch Changes
+
+- Updated dependencies [[`0832a377d`](https://github.com/bluesky-social/atproto/commit/0832a377d269584a906d5062ebb5e2e6307f9c61)]:
+  - @atproto/xrpc-server@0.7.7
+
 ## 0.1.67
 
 ### Patch Changes